rewrite pytest
